backend/requirements.txt
fastapi>=0.122.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=5.0.0
psycopg2-binary>=2.9.11
numpy>=2.0.2
scipy>=1.13.1
pydub>=0.25.1
python-dotenv>=1.0.0